Remove recordings from sparse strips renderers
