+++ sparse_strips/vello_cpu/src/dispatch/multi_threaded.rs
@@ -31,7 +31,7 @@ use vello_common::mask::Mask;
 use vello_common::paint::{ImageResolver, Paint};
 use vello_common::render_graph::RenderGraph;
 use vello_common::strip::Strip;
-use vello_common::strip_generator::{StripGenerator, StripStorage};
+use vello_common::strip_generator::StripGenerator;
 
 mod cost;
 mod worker;
@@ -99,9 +99,8 @@ pub(crate) struct MultiThreadedDispatcher {
     task_idx: u32,
     /// The number of threads active in the thread pool.
     num_threads: u16,
-    /// The strip generator for the main thread, only used for recordings.
+    /// The strip generator for the main thread, used for clip path rasterization.
     strip_generator: StripGenerator,
-    strip_storage: StripStorage,
     level: Level,
     flushed: bool,
     // So that we can reuse memory allocations across different runs.
@@ -117,13 +116,11 @@ impl MultiThreadedDispatcher {
             .num_threads(num_threads as usize)
             .build()
             .unwrap();
-        // + 1 because the main thread also stores an alpha buffer, used for recordings.
-        let alpha_storage = MaybePresent::new(vec![vec![]; usize::from(num_threads + 1)]);
+        let alpha_storage = MaybePresent::new(vec![vec![]; usize::from(num_threads)]);
         let workers = Arc::new(ThreadLocal::new());
 
         {
-            // Start counting from 1, as thread_idx 0 is reserved for the main thread.
-            let thread_ids = Arc::new(AtomicU8::new(1));
+            let thread_ids = Arc::new(AtomicU8::new(0));
             let workers = workers.clone();
 
             // Create all workers once in `new`, so that later on we can just call`.get().unwrap()`.
@@ -152,7 +149,6 @@ impl MultiThreadedDispatcher {
             task_sender: None,
             coarse_task_receiver: None,
             strip_generator: StripGenerator::new(width, height, level),
-            strip_storage: StripStorage::default(),
             level,
             alpha_storage,
             num_threads,
@@ -311,20 +307,6 @@ impl MultiThreadedDispatcher {
                             } => self.wide.generate(
                                 &task.allocation_group.strips
                                     [strip_range.start as usize..strip_range.end as usize],
-                                paint.clone(),
-                                blend_mode,
-                                thread_id,
-                                mask,
-                                encoded_paints,
-                            ),
-                            CoarseTaskType::RenderWideCommand {
-                                strips,
-                                blend_mode,
-                                paint,
-                                thread_id,
-                                mask,
-                            } => self.wide.generate(
-                                &strips,
                                 paint.clone(),
                                 blend_mode,
                                 thread_id,
@@ -560,7 +542,6 @@ impl Dispatcher for MultiThreadedDispatcher {
         self.task_sender = None;
         self.coarse_task_receiver = None;
         self.strip_generator.reset();
-        self.strip_storage.clear();
         self.alpha_storage.with_inner(|alphas| {
             for alpha in alphas {
                 alpha.clear();
@@ -595,13 +576,6 @@ impl Dispatcher for MultiThreadedDispatcher {
         // can arrive.
         drop(sender);
         self.run_coarse(false, encoded_paints);
-
-        self.alpha_storage.with_inner(|alphas| {
-            // The main thread stores the alphas that are produced by playing a recording.
-            // It is important we reserve the thread id 0 for this as the implementation for
-            // `Recordable` uses this thread ID when generating the commands for coarse rasterization.
-            alphas[0] = std::mem::take(&mut self.strip_storage.alphas);
-        });
 
         self.flushed = true;
     }
@@ -659,34 +633,6 @@ impl Dispatcher for MultiThreadedDispatcher {
         unimplemented!("composite_at_offset is not implemented for multi-threaded dispatcher");
     }
 
-    fn generate_wide_cmd(
-        &mut self,
-        strip_buf: &[Strip],
-        paint: Paint,
-        blend_mode: BlendMode,
-        _encoded_paints: &[EncodedPaint],
-    ) {
-        // Note that we are essentially round-tripping here: The wide container is inside of the
-        // main thread, but we first send a render task to a child thread which basically just
-        // forwards it back to the main thread again. We cannot apply the wide command directly
-        // here because there might be other paths that are currently being processed in a child
-        // thread that should be rendered _before_ this wide command. Therefore, we treat it like
-        // any other render task so that we can utilize the same mechanism of assigning a task ID
-        // to ensure that they are executed in order.
-        self.register_task(RenderTaskType::WideCommand {
-            strip_buf: strip_buf.into(),
-            // Recordings are currently always built on the main thread and thus have a `thread_idx`
-            // of 0.
-            thread_idx: 0,
-            paint,
-            blend_mode,
-        });
-    }
-
-    fn strip_storage_mut(&mut self) -> &mut StripStorage {
-        &mut self.strip_storage
-    }
-
     fn push_clip_path(
         &mut self,
         path: &BezPath,
@@ -707,10 +653,6 @@ impl Dispatcher for MultiThreadedDispatcher {
     fn pop_clip_path(&mut self) {
         self.flush_tasks();
         self.clip_context.pop_clip();
-    }
-
-    fn current_clip_path(&self) -> Option<vello_common::clip::PathDataRef<'_>> {
-        self.clip_context.get()
     }
 }
 
@@ -835,12 +777,6 @@ pub(crate) enum RenderTaskType {
         aliasing_threshold: Option<u8>,
         mask: Option<Mask>,
     },
-    WideCommand {
-        strip_buf: Box<[Strip]>,
-        thread_idx: u8,
-        paint: Paint,
-        blend_mode: BlendMode,
-    },
     StrokePath {
         path_range: Range<u32>,
         transform: Affine,
@@ -872,13 +808,6 @@ pub(crate) enum CoarseTaskType {
         strips: Range<u32>,
         blend_mode: BlendMode,
         paint: Paint,
-        mask: Option<Mask>,
-    },
-    RenderWideCommand {
-        thread_id: u8,
-        strips: Box<[Strip]>,
-        paint: Paint,
-        blend_mode: BlendMode,
         mask: Option<Mask>,
     },
     PushLayer {
