+++ sparse_strips/vello_hybrid/src/scene.rs
@@ -23,11 +23,8 @@ use vello_common::paint::{Paint, PaintType, Tint};
 #[cfg(feature = "text")]
 use vello_common::peniko::FontData;
 use vello_common::peniko::{BlendMode, Compose, Fill, Mix};
-use vello_common::recording::{
-    PushLayerCommand, Recordable, Recorder, Recording, RenderCommand, RenderState,
-};
+use vello_common::render_state::RenderState;
 use vello_common::render_graph::{RenderGraph, RenderNodeKind};
-use vello_common::strip::Strip;
 use vello_common::strip_generator::{GenerationMode, StripGenerator, StripStorage};
 use vello_common::util::is_axis_aligned;
 
@@ -904,259 +901,6 @@ impl Scene {
     pub fn restore_state(&mut self, state: RenderState) {
         self.render_state = state;
         self.set_paint_visible();
-    }
-}
-
-impl Recordable for Scene {
-    fn record<F>(&mut self, recording: &mut Recording, f: F)
-    where
-        F: FnOnce(&mut Recorder<'_>),
-    {
-        let mut recorder = Recorder::new(recording, self.render_state.transform);
-        f(&mut recorder);
-    }
-
-    fn prepare_recording(&mut self, recording: &mut Recording) {
-        let buffers = recording.take_cached_strips();
-        let (strip_storage, strip_start_indices) =
-            self.generate_strips_from_commands(recording.commands(), buffers);
-        recording.set_cached_strips(strip_storage, strip_start_indices);
-    }
-
-    fn execute_recording(&mut self, recording: &Recording) {
-        let (cached_strips, cached_alphas) = recording.get_cached_strips();
-        let adjusted_strips = self.prepare_cached_strips(cached_strips, cached_alphas);
-
-        // Use pre-calculated strip start indices from when we generated the cache
-        let strip_start_indices = recording.get_strip_start_indices();
-        let mut range_index = 0;
-
-        // Replay commands in order, using cached strips for geometry
-        for command in recording.commands() {
-            match command {
-                RenderCommand::FillPath(_)
-                | RenderCommand::StrokePath(_)
-                | RenderCommand::FillRect(_)
-                | RenderCommand::StrokeRect(_) => {
-                    self.process_geometry_command(
-                        strip_start_indices,
-                        range_index,
-                        &adjusted_strips,
-                    );
-                    range_index += 1;
-                }
-                RenderCommand::SetPaint(paint) => {
-                    self.set_paint(paint.clone());
-                }
-                RenderCommand::SetPaintTransform(transform) => {
-                    self.set_paint_transform(*transform);
-                }
-                RenderCommand::ResetPaintTransform => {
-                    self.reset_paint_transform();
-                }
-                RenderCommand::SetTransform(transform) => {
-                    self.set_transform(*transform);
-                }
-                RenderCommand::SetFillRule(fill_rule) => {
-                    self.set_fill_rule(*fill_rule);
-                }
-                RenderCommand::SetStroke(stroke) => {
-                    self.set_stroke(stroke.clone());
-                }
-                RenderCommand::SetTint(tint) => {
-                    self.set_tint(*tint);
-                }
-                RenderCommand::SetFilterEffect(filter) => {
-                    self.set_filter_effect(filter.clone());
-                }
-                RenderCommand::ResetFilterEffect => {
-                    self.reset_filter_effect();
-                }
-                RenderCommand::PushLayer(PushLayerCommand {
-                    clip_path,
-                    blend_mode,
-                    opacity,
-                    mask,
-                    filter,
-                }) => {
-                    self.push_layer(
-                        clip_path.as_ref(),
-                        *blend_mode,
-                        *opacity,
-                        mask.clone(),
-                        filter.clone(),
-                    );
-                }
-                RenderCommand::PopLayer => {
-                    self.pop_layer();
-                }
-            }
-        }
-    }
-}
-
-/// Recording management implementation.
-impl Scene {
-    /// Generate strips from strip commands and capture ranges.
-    ///
-    /// Returns:
-    /// - `collected_strips`: The generated strips.
-    /// - `collected_alphas`: The generated alphas.
-    /// - `strip_start_indices`: The start indices of strips for each geometry command.
-    fn generate_strips_from_commands(
-        &mut self,
-        commands: &[RenderCommand],
-        buffers: (StripStorage, Vec<usize>),
-    ) -> (StripStorage, Vec<usize>) {
-        let (mut strip_storage, mut strip_start_indices) = buffers;
-        strip_storage.clear();
-        strip_storage.set_generation_mode(GenerationMode::Append);
-        strip_start_indices.clear();
-
-        let saved_state = self.take_current_state();
-
-        for command in commands {
-            let start_index = strip_storage.strips.len();
-
-            match command {
-                RenderCommand::FillPath(path) => {
-                    self.strip_generator.generate_filled_path(
-                        path,
-                        self.render_state.fill_rule,
-                        self.render_state.transform,
-                        self.aliasing_threshold,
-                        &mut strip_storage,
-                        self.clip_context.get(),
-                    );
-                    strip_start_indices.push(start_index);
-                }
-                RenderCommand::StrokePath(path) => {
-                    self.strip_generator.generate_stroked_path(
-                        path,
-                        &self.render_state.stroke,
-                        self.render_state.transform,
-                        self.aliasing_threshold,
-                        &mut strip_storage,
-                        self.clip_context.get(),
-                    );
-                    strip_start_indices.push(start_index);
-                }
-                RenderCommand::FillRect(rect) => {
-                    self.strip_generator.generate_filled_path(
-                        rect.to_path(DEFAULT_TOLERANCE),
-                        self.render_state.fill_rule,
-                        self.render_state.transform,
-                        self.aliasing_threshold,
-                        &mut strip_storage,
-                        self.clip_context.get(),
-                    );
-                    strip_start_indices.push(start_index);
-                }
-                RenderCommand::StrokeRect(rect) => {
-                    self.strip_generator.generate_stroked_path(
-                        rect.to_path(DEFAULT_TOLERANCE),
-                        &self.render_state.stroke,
-                        self.render_state.transform,
-                        self.aliasing_threshold,
-                        &mut strip_storage,
-                        self.clip_context.get(),
-                    );
-                    strip_start_indices.push(start_index);
-                }
-                RenderCommand::SetTransform(transform) => {
-                    self.render_state.transform = *transform;
-                }
-                RenderCommand::SetFillRule(fill_rule) => {
-                    self.render_state.fill_rule = *fill_rule;
-                }
-                RenderCommand::SetStroke(stroke) => {
-                    self.render_state.stroke = stroke.clone();
-                }
-                _ => {}
-            }
-        }
-
-        self.restore_state(saved_state);
-
-        (strip_storage, strip_start_indices)
-    }
-
-    fn process_geometry_command(
-        &mut self,
-        strip_start_indices: &[usize],
-        range_index: usize,
-        adjusted_strips: &[Strip],
-    ) {
-        assert!(
-            range_index < strip_start_indices.len(),
-            "Strip range index out of bounds: range_index={}, strip_start_indices.len()={}",
-            range_index,
-            strip_start_indices.len()
-        );
-        let start = strip_start_indices[range_index];
-        let end = strip_start_indices
-            .get(range_index + 1)
-            .copied()
-            .unwrap_or(adjusted_strips.len());
-        let count = end - start;
-        if count == 0 {
-            // There are no strips to generate.
-            return;
-        }
-        assert!(
-            start < adjusted_strips.len() && count > 0,
-            "Invalid strip range: start={start}, end={end}, count={count}"
-        );
-        let paint = self.encode_current_paint();
-
-        if self.strip_path_mode != StripPathMode::CoarseOnly {
-            let mut strip_storage = self.strip_storage.borrow_mut();
-            let strip_start = strip_storage.strips.len();
-            strip_storage
-                .strips
-                .extend_from_slice(&adjusted_strips[start..end]);
-            self.fast_strips_buffer
-                .commands
-                .push(FastStripCommand::Path(FastStripsPath {
-                    strips: strip_start..strip_storage.strips.len(),
-                    paint,
-                }));
-        } else {
-            self.wide.generate(
-                &adjusted_strips[start..end],
-                paint,
-                self.render_state.blend_mode,
-                0,
-                None,
-                &self.encoded_paints.borrow(),
-            );
-        }
-    }
-
-    /// Prepare cached strips for rendering by adjusting alpha indices and extending alpha buffer.
-    #[expect(
-        clippy::cast_possible_truncation,
-        reason = "Alphas length conversion is safe in this case"
-    )]
-    fn prepare_cached_strips(
-        &mut self,
-        cached_strips: &[Strip],
-        cached_alphas: &[u8],
-    ) -> Vec<Strip> {
-        let mut strip_storage = self.strip_storage.borrow_mut();
-        // Calculate offset for alpha indices based on current buffer size.
-        let alpha_offset = strip_storage.alphas.len() as u32;
-        // Extend current alpha buffer with cached alphas.
-        strip_storage.alphas.extend(cached_alphas);
-        // Create adjusted strips with corrected alpha indices
-        cached_strips
-            .iter()
-            .map(move |strip| {
-                let mut adjusted_strip = *strip;
-                adjusted_strip.set_alpha_idx(adjusted_strip.alpha_idx() + alpha_offset);
-                adjusted_strip
-            })
-            .collect()
     }
 }
