--- sparse_strips/vello_sparse_tests/snapshots/recording_can_be_cleared.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5d3ee5faa021e9320485268cd1aabaf5cd76342643fcef97593d89f24040cc2a
-size 335
