--- sparse_strips/vello_sparse_tests/snapshots/recording_can_be_repeatedly_executed_in_layers.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:26f24dd5c60a9a7f84b039e8c19c6ddc01b9e337994e3dd3887c14067d23a54d
-size 104
