--- sparse_strips/vello_sparse_tests/snapshots/recording_handles_offscreen_content.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:108bdd50a674a7af2e73be72cefe99663d89c144aade8be489d3ab294acd01ee
-size 463
