--- sparse_strips/vello_sparse_tests/snapshots/recording_inside_clip.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:438322b7ce8c102e021ba8b532f73ccba62b5f90834d41ffd40950f76dd11fc1
-size 267
