--- sparse_strips/vello_sparse_tests/snapshots/recording_is_executed_at_recorded_transform.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:08f950f078c1ec89093ae5a9989634d458202acbe467ce6e25d388867c68cf1d
-size 95
