--- sparse_strips/vello_sparse_tests/snapshots/recording_is_executed_with_multiple_transforms.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a5bfb45d76a98c9fddfcfc8b142714d334f48ac429014ecb74fd05cbac30ad2f
-size 137
