--- sparse_strips/vello_sparse_tests/snapshots/recording_mixed_with_direct_drawing.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:891785dc2b8fb7572a2570045710c1f0e588741040aeacb3371d9e2ff3939097
-size 153
